h2 and postgres supported in flyway database migrations

+++ src/main/java/online/grigoriev/im/repository/AuditLogRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import online.grigoriev.im.entity.AuditLog;
 
 @Repository
-public interface AuditLogRepository extends JpaRepository<AuditLog, Long> {
+public interface AuditLogRepository extends JpaRepository<AuditLog, Integer> {
 }
